[UPDATE] fga-eps-mds/2018.2-GamesBI#263 Delete importdata from project
