+++ CrossData/settings.py
@@ -38,7 +38,7 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'CrossData.importdata',
+    'CrossData.API',
     'rest_framework'
 ]
